requirements.txt
pyTelegramBotAPI>=3.7.6,<4.0
python-dotenv>=0.15.0,<1.0